servicios para gestion alternativas y registro llamadas

+++ app/Http/Controllers/GestionAlternativaController.php
@@ -2,6 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use Exception;
+use App\Constants\FechaHelper;
+use App\Constants\TipoUsuario;
 use Carbon\Carbon;
 use App\Enums\MessageHttp;
 use Illuminate\Http\Request;
@@ -9,16 +12,26 @@ use App\Models\GestionAlternativa;
 use App\Services\GestionAlternativaService;
 use App\Http\Requests\StoreGestionAlternativaRequest;
 use App\Http\Requests\UpdateGestionAlternativaRequest;
+use App\Services\OrdenService;
+use App\Services\ProcuraduriaDescargaService;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Facades\Auth;
 
 class GestionAlternativaController extends Controller
 {
     protected $gestionAlternativaService;
+    protected $ordenService;
+    protected $procuraduriaDescargaService;
 
     public function __construct(
-                                 GestionAlternativaService $gestionAlternativaService
-                               )
-    {
+        GestionAlternativaService $gestionAlternativaService,
+        OrdenService $ordenService,
+        ProcuraduriaDescargaService $procuraduriaDescargaService
+    ) {
         $this->gestionAlternativaService = $gestionAlternativaService;
+        $this->ordenService = $ordenService;
+        $this->procuraduriaDescargaService = $procuraduriaDescargaService;
     }
     /**
      * Display a listing of the resource.
@@ -41,20 +54,48 @@ class GestionAlternativaController extends Controller
      */
     public function store(StoreGestionAlternativaRequest $request)
     {
-        $fechaHora = Carbon::now('America/La_Paz')->toDateTimeString();
-        $data = [
-            'solicitud_gestion' => $request->solicitud_gestion,
-            'fecha_solicitud' => $fechaHora,
-            'detalle_gestion' => '',
-            'fecha_respuesta' => null,
-            'orden_id' => $request->orden_id,
-        ];
-        $gestionAlternativa = $this->gestionAlternativaService->store($data);
 
-        return response()->json([
-            'message' => MessageHttp::CREADO_CORRECTAMENTE,
-            'data' => $gestionAlternativa
-        ], 200);
+        $fechaHora = FechaHelper::fechaHoraBolivia();
+        $descarga = $this->procuraduriaDescargaService->obtenerUnoPorOrdenId($request->orden_id);
+        if ($descarga) {
+            return response()->json([
+                'message' => 'No se puede registrar Gestión porque ya se hizo la descarga',
+                'data' => null
+            ], 409);
+        }
+        if ($this->gestionAlternativaService->hayGestionesAbiertasDeOrdenes($request->orden_id)) {
+            return response()->json([
+                'message' => 'No se puede registrar Gestión porque aún no se ha cerrado la anterior.',
+                'data' => null
+            ], 409);
+        }
+
+        DB::beginTransaction();
+        try {
+            $data = [
+                'solicitud_gestion' => $request->solicitud_gestion,
+                'fecha_solicitud' => $fechaHora,
+                'tribunal_id' => $request->tribunal_id,
+                'cuerpo_expediente_id' => $request->cuerpo_expediente_id,
+                'detalle_gestion' => '',
+                'fecha_respuesta' => null,
+                'orden_id' => $request->orden_id,
+            ];
+            $gestionAlternativa = $this->gestionAlternativaService->store($data);
+            DB::commit();
+            return response()->json([
+                'message' => MessageHttp::CREADO_CORRECTAMENTE,
+                'data' => $gestionAlternativa
+            ], 201);
+        } catch (Exception $e) {
+            DB::rollBack();
+            Log::error('Error registrar retiro: ' . $e->getMessage());
+
+            return response()->json([
+                'message' => 'Error registrar retiro',
+                'error' => $e->getMessage()
+            ], 500);
+        }
     }
 
     /**
@@ -78,18 +119,58 @@ class GestionAlternativaController extends Controller
      */
     public function update(UpdateGestionAlternativaRequest $request, GestionAlternativa $gestionAlternativa)
     {
-        $fechaHora = Carbon::now('America/La_Paz')->toDateTimeString();
-        $data = $request->only([
-            'detalle_gestion'
-        ]);
-        $data['fecha_respuesta'] = $fechaHora;
+        $tipoUsuario = Auth::user()->tipo;
+        $fechaHora = FechaHelper::fechaHoraBolivia();
+        if ($gestionAlternativa->detalle_gestion != '' && $tipoUsuario === TipoUsuario::PROCURADOR) {
+            return response()->json([
+                'message' => 'No puede hacer el registro porque el abogado ya sugirio la gestión',
+                'data' => null
+            ], 409);
+        }
+        $descarga = $this->procuraduriaDescargaService->obtenerUnoPorOrdenId($gestionAlternativa->orden_id);
+        if ($descarga) {
+            return response()->json([
+                'message' => 'No se puede registrar Gestión porque ya se hizo la descarga de la orden',
+                'data' => null
+            ], 409);
+        }
+        $cantidad = $this->gestionAlternativaService->contarGestionesPosteriores($gestionAlternativa->id, $gestionAlternativa->orden_id);
+        if ($cantidad > 0) {
+            return response()->json([
+                'message' => 'No se puede completar el registro porque ya hay una gestion posterior',
+                'data' => null
+            ], 409);
+        }
 
-        $gestionAlternativa = $this->gestionAlternativaService->update($data,$gestionAlternativa->id);
-        $data=[
-            'message'=> MessageHttp::ACTUALIZADO_CORRECTAMENTE,
-            'data'=>$gestionAlternativa
-        ];
-        return response()->json($data);
+        DB::beginTransaction();
+        try {
+
+            $data = $request->only([
+                'detalle_gestion',
+                'solicitud_gestion',
+                'tribunal_id',
+                'cuerpo_expediente_id'
+            ]);
+            if ($data['detalle_gestion']) {
+                $data['fecha_respuesta'] = $fechaHora;
+            }
+            $gestionAlternativa = $this->gestionAlternativaService->update($data, $gestionAlternativa->id);
+
+            DB::commit();
+            $data = [
+                'message' => MessageHttp::ACTUALIZADO_CORRECTAMENTE,
+                'data' => $gestionAlternativa
+            ];
+            return response()->json($data);
+        } catch (Exception $e) {
+            DB::rollBack();
+            Log::error('Error al actualizar gestion alternativa: ' . $e->getMessage());
+
+            return response()->json([
+                'message' => 'Error al actualizar gestion alternativa',
+                'error' => $e->getMessage()
+            ], 500);
+        }
     }
 
     /**
@@ -97,21 +178,56 @@ class GestionAlternativaController extends Controller
      */
     public function destroy(GestionAlternativa $gestionAlternativa)
     {
-        $gestionAlternativa = $this->gestionAlternativaService->destroy($gestionAlternativa);
-         $data=[
-            'message'=> MessageHttp::ELIMINADO_CORRECTAMENTE,
-            'data'=>$gestionAlternativa
-        ];
-        return response()->json($data);
+        if ($gestionAlternativa->detalle_gestion != '') {
+            return response()->json([
+                'message' => 'No puede anular la solicitud porque el abogado ya sugirió la gestión',
+                'data' => null
+            ], 409);
+        }
+        DB::beginTransaction();
+        try {
+            $gestionAlternativa = $this->gestionAlternativaService->destroy($gestionAlternativa);
+            DB::commit();
+            $data = [
+                'message' => MessageHttp::ELIMINADO_CORRECTAMENTE,
+                'data' => $gestionAlternativa
+            ];
+            return response()->json($data);
+        } catch (Exception $e) {
+            DB::rollBack();
+            Log::error('Error al eliminar registro: ' . $e->getMessage());
+
+            return response()->json([
+                'message' => 'Error al eliminar registro registro',
+                'error' => $e->getMessage()
+            ], 500);
+        }
     }
     public function obtenerPorOrdenId($ordenId)
     {
-        $gestionAlternativa = $this->gestionAlternativaService->obtenerPorOrdenId($ordenId);
-        $data=[
-            'message'=> MessageHttp::OBTENIDO_CORRECTAMENTE,
-            'data'=>$gestionAlternativa
+        $gestionAlternativas = $this->gestionAlternativaService->obtenerPorOrdenId($ordenId);
+        $data = [
+            'message' => MessageHttp::OBTENIDO_CORRECTAMENTE,
+            'data' => $gestionAlternativas
         ];
         return response()->json($data);
-
+    }
+    public function obtenerUnoById($gestionId)
+    {
+        $gestionAlternativas = $this->gestionAlternativaService->obtenerUnoById($gestionId);
+        $data = [
+            'message' => MessageHttp::OBTENIDO_CORRECTAMENTE,
+            'data' => $gestionAlternativas
+        ];
+        return response()->json($data);
+    }
+    public function contarGestionesPosteriores($gestionId, $ordenId)
+    {
+        $cantidad = $this->gestionAlternativaService->contarGestionesPosteriores($gestionId, $ordenId);
+        $data = [
+            'message' => MessageHttp::OBTENIDO_CORRECTAMENTE,
+            'data' => $cantidad
+        ];
+        return response()->json($data);
     }
 }
